refactor(http_exception): split into multiple files
The HttpException class and its subclasses have been refactored for better organization. The code has been divided into three separate files: http_exception_base.dart, http_4xx_exceptions.dart, and http_5xx_exceptions.dart. This change improves readability and maintainability of the code by grouping related exceptions together.
